Move some Helpers to a dedicated file
